Add ajustes na arquitetura por causa do desaclopamento do identity

+++ Service/GenericService.cs
@@ -7,7 +7,7 @@ using System.Linq.Expressions;
 
 namespace Service
 {
-    public  class GenericService<T> : IGenericService<T> where T : class
+    public class GenericService<T> : IGenericService<T> where T : class
     {
         private readonly IGenericRepository<T> _repository;
 
@@ -16,14 +16,17 @@ namespace Service
             _repository = repository;
         }
 
-        public void Delete(int id)
+        public List<string> Delete(int id)
         {
-            _repository.Delete(id);
+            if (this.Exists(id))
+                _repository.Delete(id);
+            return new List<string>();
         }
 
-        public IEnumerable<T> Get(Expression<Func<T, bool>> filter = null, Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null, string includeProperties = "")
+        public IEnumerable<T> Get(Expression<Func<T, bool>> filter = null, Expression<Func<IQueryable<T>, IOrderedQueryable<T>>> orderBy = null, string includeProperties = "")
         {
-           return  _repository.Get(filter, orderBy, includeProperties);
+
+            return _repository.Get(filter, orderBy != null ? orderBy.Compile() : null, includeProperties);
         }
 
         public IEnumerable<T> GetAll()
@@ -36,14 +39,22 @@ namespace Service
             return _repository.GetById(id);
         }
 
-        public void Insert(T obj)
+        public List<string> Insert(T obj)
         {
             _repository.Insert(obj);
+            return new List<string>();
         }
 
-        public void Update(T obj)
+        public List<string> Update(T obj)
         {
             _repository.Update(obj);
+            return new List<string>();
         }
+
+        public bool Exists(int id)
+        {
+            return (_repository.GetById(id) != null) ? true : false;
+        }
+
     }
 }
